The first part of allowing the nuclear option for students to reset assignments by deleting their submissions, just like an instructor can.
This comes with a long list of provisos, including that the assignment must be specifically opted in in the policy file, there is no feedback, it's not past the due date, and there is no grade. All except for the last one (so far anyway) I have rolled into the ACL check which we can use to prevent
a link even going back to the student, and so therefore all except for the last one are also being enforced. 

Test cases to come, plus figuring out the grade thing.



git-svn-id: https://repos.nextthought.com/svn/nti-svn/nti.contenttypes.courses/trunk@48282 3dfb626b-079f-4ab1-a9f6-4654e5377508